+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Bug Fixes**
 - Flush shall only sync the blocks to storage and not delete them from local cache.
 - Random write has been re-enabled in block cache.
+- Writing to an uncommitted block which has been deleted from the in-memory cache.
 
 ## 2.3.1 (Unreleased)
 **NOTICE**